refactor: PostItemRepository 클래스명 변경 및 패키지 이동 ( PostItemRepository -> ProductRepository )
