+++ app/src/main/java/com/mbj/ssassamarket/data/source/ProductRepository.kt
@@ -1,10 +1,11 @@
-package com.mbj.ssassamarket.data.source.remote
+package com.mbj.ssassamarket.data.source
 
 import android.util.Log
 import com.mbj.ssassamarket.data.model.ImageContent
+import com.mbj.ssassamarket.data.source.remote.MarketNetworkDataSource
 import javax.inject.Inject
 
-class PostItemRepository @Inject constructor(private val marketNetworkDataSource: MarketNetworkDataSource) {
+class ProductRepository @Inject constructor(private val marketNetworkDataSource: MarketNetworkDataSource) {
 
     suspend fun addProductPost(
         content: String,
